requirements.txt
torch>=2.11.0
torchvision>=0.26.0
transformers>=4.45.0
accelerate>=1.0.0
datasets>=3.0.0
jiwer>=3.0.0
matplotlib>=3.9.0
pandas>=2.2.0
pyyaml>=6.0.0
tqdm>=4.66.0
calflops>=0.3.0
pillow>=10.0.0
